+++ .vscode/settings.json
@@ -1,0 +1,14 @@
+{
+    "cSpell.words": [
+        "binp",
+        "deflectometry",
+        "gloo",
+        "heliostats",
+        "Juelich",
+        "nccl",
+        "NURBS",
+        "rtol",
+        "STRAL",
+        "UTIS"
+    ]
+}
