+++ artist/raytracing/rays.py
@@ -3,7 +3,7 @@ import torch
 
 class Rays:
     """
-    Model rays used for raytracing that have a direction vector and magnitude.
+    Model rays used for ray tracing that have a direction vector and magnitude.
 
     Attributes
     ----------
@@ -19,7 +19,7 @@ class Rays:
         """
         Initialize the ``Rays`` class.
 
-        The rays in ``ARTIST`` have a direction vector and a magnitude. They are used for raytracing.
+        The rays in ``ARTIST`` have a direction vector and a magnitude. They are used for ray tracing.
         The direction vector determines the direction of the rays, i.e., the path they are taking through space.
         The magnitude is important for considering atmospheric losses and cloud coverage. If a ray
         travels through a cloud, the magnitude changes.
