+++ artist/util/config_dictionary.py
@@ -15,7 +15,7 @@ target_area_curvature_e = "curvature_e"
 """Key to access the target area curvature in the east direction."""
 target_area_curvature_u = "curvature_u"
 """Key to access the target area curvature in the up direction."""
-target_area_reveicer = "receiver"
+target_area_receiver = "receiver"
 """Key used to define that the target area is a receiver."""
 target_area_type_planar = "planar"
 """Key used to define that a planar target area type is used."""
@@ -169,14 +169,14 @@ convert_nurbs_from_normals = "deflectometry"
 """Defines that the NURBS should be learnt using deflectometry data."""
 
 optimizer_use_raytracing = "use_raytracing"
-"""Defines that the alignment optimizer uses raytracing."""
+"""Defines that the alignment optimizer uses ray tracing."""
 optimizer_use_motor_positions = "use_motor_positions"
 """Defines that the alignment optimizer uses motor positions."""
 
 
 paint_target_area_geometry = "type"
 """Key to access the target area geometry in a ``PAINT`` tower.json file."""
-paint_calibration_traget = "target_name"
+paint_calibration_target = "target_name"
 """Key to access the calibration target name in a ``PAINT`` tower.json file."""
 paint_power_plant_properties = "power_plant_properties"
 """Key to access the power plant properties in a ``PAINT`` tower.json file."""
@@ -240,7 +240,7 @@ paint_number_of_facets = "number_of_facets"
 """Key to access the number of facets in a ``PAINT`` heliostat_properties.json file."""
 paint_facets = "facets"
 """Key to access the facets in a ``PAINT`` heliostat_properties.json file."""
-paint_translation_vetor = "translation_vector"
+paint_translation_vector = "translation_vector"
 """Key to access the facet translation vector in a ``PAINT`` heliostat_properties.json file.."""
 paint_canting_e = "canting_e"
 """Key to access the facet canting vector in east dimension in a ``PAINT`` heliostat_properties.json file."""
@@ -259,7 +259,7 @@ paint_utis = "UTIS"
 paint_sun_azimuth = "sun_azimuth"
 """Key to access the sun azimuth in a ``PAINT`` calibration.json file."""
 paint_sun_elevation = "sun_elevation"
-"""Key to acces the sun elevation in a ``PAINT`` calibration.json file."""
+"""Key to access the sun elevation in a ``PAINT`` calibration.json file."""
 paint_motor_positions = "motor_position"
 """Key to access motor positions in a ``PAINT`` calibration.json file."""
 paint_first_axis = "axis_1_motor_position"
